Optional formatting for Individual Study Forest Plot
For the forest plot in panel 1b for individual study results, there are now extra options to help with formatting, on screen and when downloading. 
server.R -> line 262; added format inputs into plot function.
ui.R -> line 347 (old) and 347-357 (new); added the new formatting inputs and text to explain the options and how its differs between screen and download.
fn_analysis.R -> line 161-162 (old) and 156, 161-163 (new); took text size options to be dependent on inputs within the plot function. Also made it so that the group headers were on far left for continuous data (previously it was fixed at -6).
downloadbuttons.R -> line 73-74; took text size options to be dependent on inputs for downloading.